fix test issues
