+++ test/functional-portable/corerp/cloud/resources/recipe_terraform_test.go
@@ -1,0 +1,94 @@
+/*
+Copyright 2023 The Radius Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package resource_test
+
+// This file contains tests for Terraform recipes functionality - covering general behaviors that should
+// be consistent across all resource types. These tests mostly use the extender resource type and mostly
+// avoid cloud resources to avoid unnecessary coupling and reliability issues.
+//
+// Tests in this file should only use cloud resources if absolutely necessary.
+//
+// Tests in this file should be kept *roughly* in sync with recipe_bicep_test and any other drivers.
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
+	"github.com/radius-project/radius/test/step"
+	"github.com/radius-project/radius/test/testutil"
+	"github.com/radius-project/radius/test/validation"
+)
+
+var (
+	secretNamespace = "radius-system"
+	secretPrefix    = "tfstate-default-"
+)
+
+// Test_TerraformRecipe_AzureStorage creates an Extender resource consuming a Terraform recipe that deploys an Azure blob storage instance.
+func Test_TerraformRecipe_AzureStorage(t *testing.T) {
+	template := "testdata/corerp-resources-terraform-azurestorage.bicep"
+	name := "corerp-resources-terraform-azstorage"
+	appName := "corerp-resources-terraform-azstorage-app"
+	envName := "corerp-resources-terraform-azstorage-env"
+
+	test := rp.NewRPTest(t, name, []rp.TestStep{
+		{
+			Executor: step.NewDeployExecutor(template, testutil.GetTerraformRecipeModuleServerURL(), "appName="+appName),
+			RPResources: &validation.RPResourceSet{
+				Resources: []validation.RPResource{
+					{
+						Name: envName,
+						Type: validation.EnvironmentsResource,
+					},
+					{
+						Name: appName,
+						Type: validation.ApplicationsResource,
+					},
+					{
+						Name: name,
+						Type: validation.ExtendersResource,
+						App:  appName,
+					},
+				},
+			},
+			SkipObjectValidation: true,
+			PostStepVerify: func(ctx context.Context, t *testing.T, test rp.RPTest) {
+				resourceID := "/planes/radius/local/resourcegroups/kind-radius/providers/Applications.Core/extenders/" + name
+				secretSuffix, err := corerp.GetSecretSuffix(resourceID, envName, appName)
+				require.NoError(t, err)
+
+				secret, err := test.Options.K8sClient.CoreV1().Secrets(secretNamespace).
+					Get(ctx, secretPrefix+secretSuffix, metav1.GetOptions{})
+				require.NoError(t, err)
+				require.Equal(t, secretNamespace, secret.Namespace)
+				require.Equal(t, secretPrefix+secretSuffix, secret.Name)
+			},
+		},
+	})
+
+	test.PostDeleteVerify = func(ctx context.Context, t *testing.T, test rp.RPTest) {
+		resourceID := "/planes/radius/local/resourcegroups/kind-radius/providers/Applications.Core/extenders/" + name
+		corerp.TestSecretDeletion(t, ctx, test, appName, envName, resourceID, secretNamespace, secretPrefix)
+	}
+
+	test.Test(t)
+}
