+++ test/functional-portable/corerp/noncloud/resources/testdata/corerp-azure-container-manualscale.bicep
