--- test/functional/README.md
@@ -1,3 +1,0 @@
-* [Writing functional tests](../../docs/contributing/contributing-code/contributing-code-tests/writing-functional-tests.md)
-
-* [Running functional tests](../../docs/contributing/contributing-code/contributing-code-tests/running-functional-tests.md)
